feat: centralize com styled-component
