+++ src/features/welcome/components/Centralize/style.ts
@@ -1,0 +1,9 @@
+import styled from 'styled-components/native';
+
+export const StyledView = styled.View({
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+  },
+});
